requirements-dev.txt
# Development and Setup Requirements
# These packages are needed for development, testing, and setup scripts

# Core dependencies for JIRA API interaction
requests>=2.31.0
PyYAML>=6.0.1

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Environment management
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0

# HTTP server for webhook testing
flask>=2.3.0
gunicorn>=21.2.0

# Message queue (Redis)
redis>=4.6.0

# Google Gemini API (when available)
google-generativeai>=0.3.0

# Security
cryptography>=41.0.0

# Utilities
click>=8.1.0
rich>=13.5.0
tabulate>=0.9.0
